[feat/validation] 파일 구조 도메인별 재설정
