--- src/main/kotlin/org/beyondmedicine/beyondmedicinetest/domain/constant/AccessCodeStatus.kt
@@ -1,5 +1,0 @@
-package org.beyondmedicine.beyondmedicinetest.domain.constant
-
-enum class AccessCodeStatus {
-    ACTIVE, EXPIRED
-}
