+++ src/main/kotlin/org/beyondmedicine/beyondmedicinetest/prescription/domain/constant/AccessCodeStatus.kt
@@ -1,0 +1,5 @@
+package org.beyondmedicine.beyondmedicinetest.prescription.domain.constant
+
+enum class AccessCodeStatus {
+    ACTIVE, EXPIRED
+}
