--- src/main/kotlin/org/beyondmedicine/beyondmedicinetest/repository/accesscode/UserAccessCodeRepository.kt
@@ -1,11 +1,0 @@
-package org.beyondmedicine.beyondmedicinetest.repository.accesscode
-
-import org.beyondmedicine.beyondmedicinetest.domain.accesscode.UserAccessCode
-import org.beyondmedicine.beyondmedicinetest.domain.constant.AccessCodeStatus
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface UserAccessCodeRepository: JpaRepository<UserAccessCode, Long> {
-
-    fun findByUserIdAndStatus(userId: String, status: AccessCodeStatus): UserAccessCode?
-
-}
